Moved local-vol FD operator to core library.
It was already used by core classes, and the interface is stable.
