+++ QuantLib.vcxproj
@@ -688,7 +688,6 @@
     <ClInclude Include="ql\experimental\finitedifferences\fdmhestongreensfct.hpp" />
     <ClInclude Include="ql\experimental\finitedifferences\fdmklugeextouop.hpp" />
     <ClInclude Include="ql\experimental\finitedifferences\fdmklugeextousolver.hpp" />
-    <ClInclude Include="ql\experimental\finitedifferences\fdmlocalvolfwdop.hpp" />
     <ClInclude Include="ql\experimental\finitedifferences\fdmsimple2dextousolver.hpp" />
     <ClInclude Include="ql\experimental\finitedifferences\fdmsimple3dextoujumpsolver.hpp" />
     <ClInclude Include="ql\experimental\finitedifferences\fdmspreadpayoffinnervalue.hpp" />
@@ -1163,6 +1162,7 @@
     <ClInclude Include="ql\methods\finitedifferences\operators\fdmlinearopcomposite.hpp" />
     <ClInclude Include="ql\methods\finitedifferences\operators\fdmlinearopiterator.hpp" />
     <ClInclude Include="ql\methods\finitedifferences\operators\fdmlinearoplayout.hpp" />
+    <ClInclude Include="ql\methods\finitedifferences\operators\fdmlocalvolfwdop.hpp" />
     <ClInclude Include="ql\methods\finitedifferences\operators\fdmornsteinuhlenbeckop.hpp" />
     <ClInclude Include="ql\methods\finitedifferences\operators\fdmsabrop.hpp" />
     <ClInclude Include="ql\methods\finitedifferences\operators\firstderivativeop.hpp" />
@@ -1988,7 +1988,6 @@
     <ClCompile Include="ql\experimental\finitedifferences\fdmhestonfwdop.cpp" />
     <ClCompile Include="ql\experimental\finitedifferences\fdmhestongreensfct.cpp" />
     <ClCompile Include="ql\experimental\finitedifferences\fdmklugeextouop.cpp" />
-    <ClCompile Include="ql\experimental\finitedifferences\fdmlocalvolfwdop.cpp" />
     <ClCompile Include="ql\experimental\finitedifferences\fdmsquarerootfwdop.cpp" />
     <ClCompile Include="ql\experimental\finitedifferences\fdmvppstartlimitstepcondition.cpp" />
     <ClCompile Include="ql\experimental\finitedifferences\fdmvppstepcondition.cpp" />
@@ -2270,6 +2269,7 @@
     <ClCompile Include="ql\methods\finitedifferences\operators\fdmhestonop.cpp" />
     <ClCompile Include="ql\methods\finitedifferences\operators\fdmhullwhiteop.cpp" />
     <ClCompile Include="ql\methods\finitedifferences\operators\fdmlinearoplayout.cpp" />
+    <ClCompile Include="ql\methods\finitedifferences\operators\fdmlocalvolfwdop.cpp" />
     <ClCompile Include="ql\methods\finitedifferences\operators\fdmornsteinuhlenbeckop.cpp" />
     <ClCompile Include="ql\methods\finitedifferences\operators\fdmsabrop.cpp" />
     <ClCompile Include="ql\methods\finitedifferences\operators\firstderivativeop.cpp" />
